requirements.txt
# Core dependencies
langchain>=0.1.0,<1.0.0
langgraph>=0.0.40
openai>=1.3.0
neo4j>=5.0.0
tiktoken>=0.5.0

# FastAPI and API dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Document processing
pypdf>=3.17.0
python-docx>=0.8.11
python-multipart>=0.0.6
python-pptx>=0.6.21
openpyxl>=3.1.0

# OCR and image processing
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.17.0
opencv-python>=4.8.0
langdetect>=1.0.9

# Vector embeddings
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# Configuration and environment
pydantic>=2.0.0
pydantic-settings>=1.0.0
python-dotenv>=1.0.0

# Caching
cachetools==5.3.2
diskcache>=5.6.0



# Utilities
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
aiofiles>=23.0.0
python-igraph>=0.11.0
leidenalg>=0.10.1
reportlab>=4.0.0

# Graph visualization (optional - not used in new UI)
plotly>=5.17.0
networkx>=3.2.0

# Logging and monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development dependencies
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# Testcontainers for e2e tests (starts Neo4j programmatically)
testcontainers>=3.11.0

# Optional: Jupyter for development
jupyter>=1.0.0

# Redis for job queue (optional)
redis>=4.6.0

# FlashRank reranker (optional)
flashrank

# OpenTelemetry distributed tracing (optional)
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-instrumentation>=0.41b0
